Update: zsh script

+++ install_zsh.sh
@@ -9,13 +9,17 @@ try_install /bin/zsh zsh
 
 # --- install oh my zsh
 
-cd ${HOME} && sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [ ! -e "${HOME}/.oh-my-zsh" ]; then
+    log "oh-my-zsh not found, installing..."
+    cd ${HOME} && sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
-if [ $? -ne 0 ]; then
-    error "Oh my zsh installation failed."
-    exit 1
-else
-    log "Oh my zsh installation succeeded."
+    if [ $? -ne 0 ]; then
+        error "Oh my zsh installation failed."
+        exit 1
+    else
+        log "Oh my zsh installation succeeded."
+    fi
+fi
 
 # --- install autojump
 
@@ -23,15 +27,20 @@ try_install /usr/bin/autojump autojump
 
 # --- install zsh suggestions
 
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-if [ $? -ne 0 ]; then
-    error "Zsh-autosuggestions installation failed."
-    exit 1
-else
-    log "Zsh-autosuggestions installation succeeded."
+if [ ! -e "${HOME}/.oh-my-zsh/custom/plugins/zsh-autosuggestions" ]; then
+    log "zsh-autosuggestions not found, installing..."
+    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+    
+    if [ $? -ne 0 ]; then
+        error "Zsh-autosuggestions installation failed."
+        exit 1
+    else
+        log "Zsh-autosuggestions installation succeeded."
+    fi
 fi
 
 # --- install chainy zsh theme
+scp `pwd`/.zsh/.zshrc ${HOME}/
 scp `pwd`/.zsh/chainy-zsh-theme/chainy.zsh-theme ${HOME}/.oh-my-zsh/themes/
 
 # --- change shell
